pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ogc-api-records-core-client"
dynamic = ["version"]
description = "A client library for accessing Invenio REST API"

readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Fabrice Brito", email = "info@terradue.com" },
  { name = "Simone Tripodi", email = "info@terradue.com"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "httpx>=0.23.0,<0.29.0",
  "attrs>=22.2.0",
  "python-dateutil==2.8.0",
  "loguru"
]

[tool.hatch.metadata]
allow-direct-references = true

[project.urls]
Documentation = "https://github.com/Terradue/ogc-api-records-core-client#readme"
Issues = "https://github.com/Terradue/ogc-api-records-core-client/issues"
Source = "https://github.com/Terradue/ogc-api-records-core-client"

[tool.hatch.version]
path = "ogc_api_records_core_client/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:ogc_api_records_core_client tests}"

[tool.coverage.run]
source_pkgs = ["ogc_api_records_core_client", "tests"]
branch = true
parallel = true
omit = [
  "ogc_api_records_core_client/__about__.py",
]

[tool.coverage.paths]
ogc_api_records_core_client = ["ogc_api_records_core_client", "*/ogc_api_records_core_client"]
tests = ["tests", "*/ogc_api_records_core_client/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.envs.test]
skip-install = false

dependencies = [
  "httpx>=0.23.0,<0.29.0",
  "attrs>=22.2.0",
  "python-dateutil==2.8.0",
  "loguru"
]

[tool.hatch.envs.test.scripts]
test = "hatch run nose2"
testv = "hatch run nose2 --verbose"
cov = ["coverage run --source=ogc_api_records_core_client -m nose2", "coverage report"]

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12", "3.13"]
